Make the `redis` network "internal"
I'm not sure why thomasjpfan has `external: true` for the `redis` network. It's not obvious why it's needed, and we'd almost certainly prefer to have the network "private" so people can't poke the redis components from the outside.

I'm hoping that specifying in `docker-compose.yml` that the `redis` network is
an overlay network and attachable will allow the various hosts and containers to connect to it. Not at all certain, though.