Copy `stackdemo` source over
This copies the `stackdemo` source from https://github.com/collabnix/dockerlabs/tree/master/solution/redis/viz-web-redis/stackdemo into this project.

I needed to do this to modify the name of the Redis host. Referring to `redis` tended to fail because it was trying to poke a slave instead of the master. I'm hoping that changing the name to `redismaster` will fix that, but to see I needed to build my own image.