Attempt to fix problem accessing Redis
The Flask clients weren't successfully contacting a Redis client, sometimes at all, and other times because they'd get a "slave" when they'd need a "master". I think this might fix that?

I also added a second route, so we can just get the count without updating it, and increment and return the count. The response now includes the name and ID, although I'm not sure _what_ they're the name and ID of, and whether they will be useful.